[refactor] Change calcLengthScore return type from int to double
int truncation silently loses precision for cases where the formula
produces a non-integer result (e.g. lengths 7 vs 11 yields 25.714...).
Return double to preserve full precision of the score.

Add NonIntegerResult_RetainsPrecision test to cover this case.

+++ simchecker.cpp
@@ -1,13 +1,13 @@
 #include "simchecker.h"
 #include <algorithm>
 
-int SimChecker::calcLengthScore(const std::string& a, const std::string& b) const {
+double SimChecker::calcLengthScore(const std::string& a, const std::string& b) const {
     int longer  = static_cast<int>(std::max(a.size(), b.size()));
     int shorter = static_cast<int>(std::min(a.size(), b.size()));
 
     if (longer >= 2 * shorter)
-        return 0;
+        return 0.0;
 
     int gap = longer - shorter;
-    return static_cast<int>((1.0 - static_cast<double>(gap) / shorter) * 60);
+    return (1.0 - static_cast<double>(gap) / shorter) * 60;
 }
